+++ server/src/modes/style/stylus/index.ts
@@ -37,7 +37,12 @@ export function getStylusMode(
     doComplete(document, position) {
       const embedded = embeddedDocuments.refreshAndGet(document);
 
-      const lsCompletions = provideCompletionItems(embedded, position);
+      const useSeparator =
+        typeof env.getConfig().languageStylus.useSeparator === 'undefined'
+          ? true
+          : env.getConfig().languageStylus.useSeparator;
+
+      const lsCompletions = provideCompletionItems(embedded, position, useSeparator);
       const lsItems = _.map(lsCompletions.items, i => {
         return {
           ...i,
@@ -74,8 +79,10 @@ export function getStylusMode(
         return [];
       }
 
-      const stylusSupremacy: IStylusSupremacy = dependencyService.get('stylus-supremacy', getFileFsPath(document.uri))
-        .module;
+      const stylusSupremacy: IStylusSupremacy = dependencyService.get(
+        'stylus-supremacy',
+        getFileFsPath(document.uri)
+      ).module;
 
       const inputText = document.getText(range);