+++ vti/src/initParams.ts
@@ -89,6 +89,7 @@ function getDefaultVLSConfig() {
       format: {}
     },
     emmet: {},
-    stylusSupremacy: {}
+    stylusSupremacy: {},
+    languageStylus: {}
   };
 }